// src/app/generated/player-api/model/eventType.ts
/**
 * Player API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const EventType = {
    ViewCreated: 'ViewCreated',
    ViewDeleted: 'ViewDeleted'
} as const;
export type EventType = typeof EventType[keyof typeof EventType];