// src/app/generated/player-api/model/webhookSubscriptionForm.ts
/**
 * Player API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventType } from './eventType';


export interface WebhookSubscriptionForm { 
    name?: string | null;
    callbackUri?: string | null;
    clientId?: string | null;
    clientSecret?: string | null;
    eventTypes?: Array<EventType> | null;
}